# net/http/httputil: use io.CopyBuffer instead of custom copyBuffer