Remove the duplicate implementation of copybuffer and simplify the code.
go  version
go version go1.15 linux/amd64

 Here is the benchstat result:

/root/go/bin/benchstat after.txt before.txt 
name         old time/op    new time/op    delta
ServeHTTP-8    11.2µs ± 4%    11.0µs ± 4%   ~     (p=0.278 n=10+9)

name         old alloc/op   new alloc/op   delta
ServeHTTP-8    33.6kB ± 0%    33.6kB ± 0%   ~     (all equal)

name         old allocs/op  new allocs/op  delta
ServeHTTP-8      6.00 ± 0%      6.00 ± 0%   ~     (all equal)